+++ lib/admiralui_flutter.dart
@@ -238,6 +238,7 @@ export 'package:admiralui_flutter/src/widgets/views/toolbar/toolbar_widget.dart'
 export 'package:admiralui_flutter/src/widgets/views/toolbar/toolbar_widget_scheme.dart';
 export 'package:admiralui_flutter/src/widgets/views/zero_screen/zero_screen.dart';
 export 'package:admiralui_flutter/src/widgets/views/zero_screen/zero_screen_scheme.dart';
+export 'package:admiralui_flutter/src/widgets/widget_size/child_size_notifier.dart';
 export 'package:admiralui_flutter/src/widgets/widget_size/widget_size.dart';
 export 'package:admiralui_flutter/theme/admiral_font.dart';
 export 'package:admiralui_flutter/theme/app_theme.dart';
